lib/pty: Ensure waitpid is repeated after EINTR
The waitpid syscall may return EINTR. In that case, the waitpid call
should be repeated to ensure that children are reaped.

I have seen a case of script not reaping its child. While I have not
been able to verify that this caused the issue, it does seem like a
likely cause for the problem.

Signed-off-by: Benjamin Berg <bberg@redhat.com>