uuidd: move option parsing to separate function
This allows daemon to free option parsing related stack memory when it is no
longer useful.  While doing that move also buffer variable to context where
it is needed, and avoid using extensively large buffer when there is no need
for it.

Finally change atoi() to strtou32_or_err() when parsing a pid to catch what
should be impossible - uuidd returning not a number.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>